Refactoring version specific stuff
